requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic
pytest
httpx
pytest-asyncio
pydantic[email]
python-multipart